Style shoutbox messages
