+++ intranet-client/src/panels/shoutbox/ShoutboxPanel.tsx
@@ -1,30 +1,8 @@
 import * as React from 'react';
 
 import Panel, { PanelProps } from '../Panel';
-import useShoutbox, { Message } from './useShoutbox';
-
-interface ShoutboxMessageProps {
-    message: Message;
-}
-
-function ShoutboxMessage({ message }: ShoutboxMessageProps) {
-    return (
-        <article className="ShoutboxMessage">
-            <header className="ShoutboxMessage__header">
-                <span className="ShoutboxMessage__username">
-                    {message.username}
-                </span>
-                <time className="ShoutboxMessage__time">
-                    {message.time.toLocaleString('nl-NL', {
-                        hour: '2-digit',
-                        minute: '2-digit',
-                    })}
-                </time>
-            </header>
-            <p className="ShoutboxMessage__body">{message.body}</p>
-        </article>
-    );
-}
+import { ShoutboxMessage } from './ShoutboxMessage';
+import useShoutbox from './useShoutbox';
 
 export type ShoutboxPanelProps = PanelProps;
